src: move Util to the util folder
