+++ src/components/Navbar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import NavOption from './NavOption';
-import { redirect } from '../Util';
+import { redirect } from '../util/Util';
 
 const Navbar = () => (
   <nav className="navbar" role="navigation" aria-label="main navigation">
